♻️ Refactor: Controller 리팩토링
generateLottos, calculateProfitRate -> Controller.LottoController 파일 내 generateLottos, calculateProfitRate 함수로 변경
