--- src/main/kotlin/lotto/Controller/CalculateProfitRate.kt
@@ -1,8 +1,0 @@
-package lotto.Controller
-
-import lotto.Model.Prize
-
-fun calculateProfitRate(results: Map<Prize, Int>, purchaseAmount: Int): Double {
-    val totalPrizeMoney = results.entries.sumOf { (prize, count) -> prize.prizeMoney * count }
-    return (totalPrizeMoney / purchaseAmount.toDouble()) * 100
-}
