+++ src/main/kotlin/lotto/Controller/LottoController.kt
@@ -1,5 +1,6 @@
 package lotto.Controller
 
+import camp.nextstep.edu.missionutils.Randoms
 import lotto.Model.*
 import lotto.View.*
 
@@ -18,5 +19,17 @@ class LottoController {
         val profitRate = calculateProfitRate(results, purchaseAmount)
 
         printResults(results, profitRate)
+    }
+
+    private fun generateLottos(count: Int): List<Lotto> {
+        return List(count) {
+            val numbers = Randoms.pickUniqueNumbersInRange(1, 45, 6).sorted()
+            Lotto(numbers)
+        }
+    }
+
+    private fun calculateProfitRate(results: Map<Prize, Int>, purchaseAmount: Int): Double {
+        val totalPrizeMoney = results.entries.sumOf { (prize, count) -> prize.prizeMoney * count }
+        return (totalPrizeMoney / purchaseAmount.toDouble()) * 100
     }
 }
